requirements.txt
# Tested with Python 3.12 (conda env: azcli)
# Standard install:
#   pip install -r requirements.txt
#
# NOTE: svglib requires system cairo libraries.
# If pycairo fails to compile, install svglib via conda instead:
#   conda install -c conda-forge svglib

reportlab>=4.4.4
PyMuPDF>=1.26.5
pandas>=2.3.0
openpyxl>=3.1.5
python-dotenv>=1.1.1
Pillow>=12.0.0
qrcode[pil]>=8.2
svglib>=1.5.1